+++ src/main/java/com/raoulvdberge/refinedpipes/RefinedPipes.java
@@ -1,5 +1,6 @@
 package com.raoulvdberge.refinedpipes;
 
+import com.raoulvdberge.refinedpipes.config.ServerConfig;
 import com.raoulvdberge.refinedpipes.item.group.MainItemGroup;
 import com.raoulvdberge.refinedpipes.setup.ClientSetup;
 import com.raoulvdberge.refinedpipes.setup.CommonSetup;
@@ -10,7 +11,9 @@ import net.minecraft.tileentity.TileEntityType;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.DistExecutor;
+import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 @Mod(RefinedPipes.ID)
@@ -18,11 +21,14 @@ public class RefinedPipes {
     public static final String ID = "refinedpipes";
     public static final ItemGroup MAIN_GROUP = new MainItemGroup();
     public static final RefinedPipesNetwork NETWORK = new RefinedPipesNetwork();
+    public static final ServerConfig SERVER_CONFIG = new ServerConfig();
 
     public RefinedPipes() {
         CommonSetup commonSetup = new CommonSetup();
 
         DistExecutor.runWhenOn(Dist.CLIENT, () -> ClientSetup::new);
+
+        ModLoadingContext.get().registerConfig(ModConfig.Type.SERVER, SERVER_CONFIG.getSpec());
 
         FMLJavaModLoadingContext.get().getModEventBus().addGenericListener(Block.class, commonSetup::onRegisterBlocks);
         FMLJavaModLoadingContext.get().getModEventBus().addGenericListener(Item.class, commonSetup::onRegisterItems);